separated UI and logic
