+++ lib/screens/home_screen.dart
@@ -1,14 +1,15 @@
 import 'dart:io';
 
 import 'package:all_social_app/models/users.dart';
+import 'package:all_social_app/screens/create_posts/frame_select_screen.dart';
 import 'package:all_social_app/widgets/bottom_navbar.dart';
+import 'package:all_social_app/widgets/no_posts.dart';
+import 'package:all_social_app/widgets/show_posts.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:google_fonts/google_fonts.dart';
-
-import 'create_posts/step_1_screen.dart';
 
 class HomeScreen extends StatefulWidget {
   final String? currentDocId;
@@ -63,6 +64,7 @@ class _HomeWidgetState extends State<HomeWidget> {
     ),
   );
   int hours = DateTime.now().hour;
+
   late String docId = "docSnapshot.id";
   Future<UserFire?> readUser() async {
     await FirebaseFirestore.instance
@@ -79,7 +81,6 @@ class _HomeWidgetState extends State<HomeWidget> {
     );
 
     final docUser = FirebaseFirestore.instance.collection('users').doc(docId);
-
     final snapshot = await docUser.get();
 
     if (snapshot.exists) {
@@ -104,73 +105,7 @@ class _HomeWidgetState extends State<HomeWidget> {
     if (posts.isEmpty) {
       return NoPosts(widget: widget);
     } else {
-      return Column(
-        children: posts.map((post) {
-          return ListTile(
-            title: Padding(
-              padding: const EdgeInsets.only(bottom: 5),
-              child: Container(
-                height: 390,
-                width: 342,
-                color: const Color(0xffFCE6EE),
-                child: Padding(
-                  padding: const EdgeInsets.all(10),
-                  child: Column(
-                    crossAxisAlignment: CrossAxisAlignment.center,
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Container(
-                        height: 322,
-                        width: 322,
-                        decoration: BoxDecoration(
-                          borderRadius: BorderRadius.circular(4),
-                        ),
-                        child: Padding(
-                          padding: const EdgeInsets.only(bottom: 10),
-                          child: ClipRRect(
-                            borderRadius: BorderRadius.circular(4.0),
-                            child: Image.network(
-                              post.post,
-                              fit: BoxFit.fill,
-                            ),
-                          ),
-                        ),
-                      ),
-                      Padding(
-                        padding: const EdgeInsets.symmetric(horizontal: 6),
-                        child: Row(
-                          mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          children: [
-                            Column(
-                              mainAxisAlignment: MainAxisAlignment.start,
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Text(
-                                  "Schedule: ${post.postDate}",
-                                  style: textstyle,
-                                ),
-                                Text(
-                                  post.postPlatform,
-                                  style: textstyle,
-                                ),
-                              ],
-                            ),
-                            Text(
-                              post.postTime,
-                              style: textstyle,
-                            ),
-                          ],
-                        ),
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ),
-          );
-        }).toList(),
-      );
+      return ShowPosts(posts: posts, textstyle: textstyle);
     }
   }
 
@@ -178,128 +113,89 @@ class _HomeWidgetState extends State<HomeWidget> {
   Widget build(BuildContext context) {
     return Stack(
       children: [
-        SingleChildScrollView(
-          child:
-              // FutureBuilder(
-              //     future: readUser(),
-              //     builder: (context, snapshot) {
-              //       if (snapshot.hasData) {
-              //         final user = snapshot.data;
-              //         return Text(user!.userName);
-              //       }
-              //       return const CircularProgressIndicator();
-              //     }),
-              FutureBuilder(
-            future: readUser(),
-            builder: (context, snapshot) {
-              if (snapshot.connectionState == ConnectionState.waiting) {
-                return const Center(
-                  child: CircularProgressIndicator(),
-                );
-              } else if (snapshot.hasError) {
-                return Center(child: Text('Error: ${snapshot.error}'));
-              } else if (snapshot.hasData) {
-                final userName = snapshot.data?.userName;
-                final image = snapshot.data?.profileImage;
+        FutureBuilder(
+          future: readUser(),
+          builder: (context, snapshot) {
+            if (snapshot.connectionState == ConnectionState.waiting) {
+              return const Center(
+                child: CircularProgressIndicator(),
+              );
+            } else if (snapshot.hasError) {
+              return Center(
+                child: Text('Error: ${snapshot.error}'),
+              );
+            } else if (snapshot.hasData) {
+              final userName = snapshot.data?.userName;
+              final image = snapshot.data?.profileImage;
 
-                return Column(
-                  mainAxisSize: MainAxisSize.max,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.fromLTRB(24, 48, 24, 0),
-                      child: Row(
-                        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                        children: [
-                          Text(
-                            hours >= 1 && hours <= 12
-                                ? 'Good Morning!\n$userName'
-                                : hours >= 12 && hours <= 16
-                                    ? 'Good Afternoon!\n$userName'
-                                    : hours >= 16 && hours <= 21
-                                        ? 'Good Evening!\n$userName'
-                                        : 'Good Night!\n$userName',
-                            style: GoogleFonts.montserrat(
-                              textStyle: const TextStyle(
-                                fontWeight: FontWeight.w500,
-                                fontSize: 24,
-                                color: Color(0xff1C1C1C),
-                              ),
+              return Column(
+                mainAxisSize: MainAxisSize.max,
+                children: [
+                  Padding(
+                    padding: const EdgeInsets.fromLTRB(24, 48, 24, 0),
+                    child: Row(
+                      mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                      children: [
+                        Text(
+                          hours >= 1 && hours <= 12
+                              ? 'Good Morning!\n$userName'
+                              : hours >= 12 && hours <= 16
+                                  ? 'Good Afternoon!\n$userName'
+                                  : hours >= 16 && hours <= 21
+                                      ? 'Good Evening!\n$userName'
+                                      : 'Good Night!\n$userName',
+                          style: GoogleFonts.montserrat(
+                            textStyle: const TextStyle(
+                              fontWeight: FontWeight.w500,
+                              fontSize: 24,
+                              color: Color(0xff1C1C1C),
                             ),
                           ),
-                          Container(
-                            height: 46,
-                            width: 46,
-                            decoration: const BoxDecoration(
-                              shape: BoxShape.circle,
-                            ),
-                            child: ClipRRect(
-                              borderRadius: BorderRadius.circular(300.0),
-                              child: Image.file(
-                                File(image!),
-                                fit: BoxFit.fill,
-                                height: 100,
-                                width: 100,
-                              ),
+                        ),
+                        Container(
+                          height: 46,
+                          width: 46,
+                          decoration: const BoxDecoration(
+                            shape: BoxShape.circle,
+                          ),
+                          child: ClipRRect(
+                            borderRadius: BorderRadius.circular(300.0),
+                            child: Image.file(
+                              File(image!),
+                              fit: BoxFit.fill,
+                              height: 100,
+                              width: 100,
                             ),
                           ),
-                        ],
-                      ),
+                        ),
+                      ],
                     ),
-                    FutureBuilder(
-                      future: _buildPosts(),
-                      builder: (BuildContext context,
-                          AsyncSnapshot<Widget> snapshot) {
-                        if (snapshot.connectionState ==
-                            ConnectionState.waiting) {
-                          return const Center(
-                              child: CircularProgressIndicator());
-                        } else if (snapshot.hasError) {
-                          return Center(
-                              child: Text('Error: ${snapshot.error}'));
-                        } else {
-                          return snapshot.data!;
-                        }
-                      },
-                    ),
-                    // FutureBuilder<List<Posts>>(
-                    //   future: _postsFuture,
-                    //   builder: (BuildContext context,
-                    //       AsyncSnapshot<List<Posts>> snapshot) {
-                    //     if (snapshot.connectionState ==
-                    //         ConnectionState.waiting) {
-                    //       return const Center(
-                    //         child: CircularProgressIndicator(),
-                    //       );
-                    //     } else if (snapshot.hasError) {
-                    //       return Center(
-                    //         child: Text('Error: ${snapshot.error}'),
-                    //       );
-                    //     } else if (snapshot.hasData) {
-                    //       final List<Posts>? posts = snapshot.data;
-                    //       if (posts!.isEmpty) {
-                    //         return NoPosts(widget: widget);
-                    //       } else {
-                    //         return
-                    //          ShowPosts(posts: posts);
-                    //       }
-                    //     } else {
-                    //       return Container(
-                    //         color: Colors.red,
-                    //         height: 550,
-                    //         width: 500,
-                    //       );
-                    //     }
-                    //   },
-                    // ),
-                  ],
-                );
-              } else {
-                return const Center(
-                  child: Text('no data'),
-                );
-              }
-            },
-          ),
+                  ),
+                  FutureBuilder(
+                    future: _buildPosts(),
+                    builder:
+                        (BuildContext context, AsyncSnapshot<Widget> snapshot) {
+                      if (snapshot.connectionState == ConnectionState.waiting) {
+                        return const Center(
+                          child: Text(''),
+                        );
+                      } else if (snapshot.hasError) {
+                        return Center(
+                          child: Text('Error: ${snapshot.error}'),
+                        );
+                      } else {
+                        return snapshot.data!;
+                      }
+                    },
+                  ),
+                ],
+              );
+            } else {
+              return const Center(
+                child: Text('no data'),
+              );
+            }
+          },
         ),
         Positioned(
           bottom: 6,
@@ -313,7 +209,7 @@ class _HomeWidgetState extends State<HomeWidget> {
                     Navigator.push(
                       context,
                       MaterialPageRoute(
-                        builder: (context) => CreatePostScreenStep1(),
+                        builder: (context) => const FrameSelectScreen(),
                       ),
                     );
                   },
@@ -325,80 +221,6 @@ class _HomeWidgetState extends State<HomeWidget> {
                   color: Colors.transparent,
                 ),
         )
-      ],
-    );
-  }
-}
-
-class NoPosts extends StatelessWidget {
-  const NoPosts({
-    super.key,
-    required this.widget,
-  });
-
-  final HomeWidget widget;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        Padding(
-          padding: const EdgeInsets.fromLTRB(24, 55, 24, 0),
-          child: SvgPicture.asset(
-            "assets/no_posts_default_image.svg",
-            height: 342,
-            width: 342,
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(0, 30, 0, 50),
-          child: Text(
-            "You don’t have create any posts. Please\ncreate a new post.",
-            textAlign: TextAlign.center,
-            style: GoogleFonts.montserrat(
-              textStyle: const TextStyle(
-                fontWeight: FontWeight.w400,
-                fontSize: 16,
-                color: Color(0xff1C1C1C),
-              ),
-            ),
-          ),
-        ),
-        Padding(
-          padding: const EdgeInsets.fromLTRB(0, 0, 0, 10),
-          child: GestureDetector(
-            onTap: () {
-              Navigator.push(
-                context,
-                MaterialPageRoute(
-                  builder: (context) => CreatePostScreenStep1(),
-                ),
-              );
-            },
-            child: Container(
-              height: 40,
-              width: 342,
-              decoration: const BoxDecoration(
-                borderRadius: BorderRadius.all(
-                  Radius.circular(6),
-                ),
-                color: Color(0xffED4D86),
-              ),
-              child: Center(
-                child: Text(
-                  'Create Post',
-                  style: GoogleFonts.montserrat(
-                    textStyle: const TextStyle(
-                      fontWeight: FontWeight.w600,
-                      fontSize: 16,
-                      color: Color(0xffFFFFFC),
-                    ),
-                  ),
-                ),
-              ),
-            ),
-          ),
-        ),
       ],
     );
   }
