+++ lib/widgets/onboarding/dot_indicator.dart
@@ -1,0 +1,24 @@
+import 'package:flutter/material.dart';
+
+class DotIndicator extends StatelessWidget {
+  const DotIndicator({
+    super.key,
+    this.isActive = false,
+  });
+
+  final bool isActive;
+
+  @override
+  Widget build(BuildContext context) {
+    return Container(
+      height: 12,
+      width: 12,
+      decoration: BoxDecoration(
+        color: isActive ? const Color(0xffED4D86) : const Color(0xffFCE6EE),
+        borderRadius: const BorderRadius.all(
+          Radius.circular(100),
+        ),
+      ),
+    );
+  }
+}
