+++ lib/widgets/show_posts.dart
@@ -3,26 +3,21 @@ import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 
 class ShowPosts extends StatelessWidget {
-  ShowPosts({
+  const ShowPosts({
     super.key,
     required this.posts,
+    required this.textstyle,
   });
 
-  final List<Posts>? posts;
-
-  TextStyle textstyle = GoogleFonts.montserrat(
-    textStyle: const TextStyle(
-      fontWeight: FontWeight.w400,
-      fontSize: 14,
-    ),
-  );
+  final List<PostFire> posts;
+  final TextStyle textstyle;
 
   @override
   Widget build(BuildContext context) {
     return Column(
       children: [
         Padding(
-          padding: const EdgeInsets.fromLTRB(24, 24, 24, 0),
+          padding: const EdgeInsets.fromLTRB(24, 24, 24, 16),
           child: Align(
             alignment: Alignment.centerLeft,
             child: Text(
@@ -37,74 +32,71 @@ class ShowPosts extends StatelessWidget {
           ),
         ),
         Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: posts!.map<Widget>(
-            (Posts post) {
-              return ListTile(
-                title: Padding(
-                  padding: const EdgeInsets.only(bottom: 5),
-                  child: Container(
-                    height: 390,
-                    width: 342,
-                    color: const Color(0xffFCE6EE),
-                    child: Padding(
-                      padding: const EdgeInsets.all(10),
-                      child: Column(
-                        crossAxisAlignment: CrossAxisAlignment.center,
-                        mainAxisAlignment: MainAxisAlignment.center,
-                        children: [
-                          Container(
-                            height: 322,
-                            width: 322,
-                            decoration: BoxDecoration(
-                              borderRadius: BorderRadius.circular(4),
-                            ),
-                            child: Padding(
-                              padding: const EdgeInsets.only(bottom: 10),
-                              child: ClipRRect(
-                                borderRadius: BorderRadius.circular(4.0),
-                                child: Image.memory(
-                                  post.post,
-                                  fit: BoxFit.fill,
-                                ),
+          children: posts.map((post) {
+            return ListTile(
+              title: Padding(
+                padding: const EdgeInsets.only(bottom: 5),
+                child: Container(
+                  height: 390,
+                  width: 342,
+                  color: const Color(0xffFCE6EE),
+                  child: Padding(
+                    padding: const EdgeInsets.all(10),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.center,
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: [
+                        Container(
+                          height: 322,
+                          width: 322,
+                          decoration: BoxDecoration(
+                            borderRadius: BorderRadius.circular(4),
+                          ),
+                          child: Padding(
+                            padding: const EdgeInsets.only(bottom: 10),
+                            child: ClipRRect(
+                              borderRadius: BorderRadius.circular(4.0),
+                              child: Image.network(
+                                post.post,
+                                fit: BoxFit.fill,
                               ),
                             ),
                           ),
-                          Padding(
-                            padding: const EdgeInsets.symmetric(horizontal: 6),
-                            child: Row(
-                              mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                              crossAxisAlignment: CrossAxisAlignment.start,
-                              children: [
-                                Column(
-                                  mainAxisAlignment: MainAxisAlignment.start,
-                                  crossAxisAlignment: CrossAxisAlignment.start,
-                                  children: [
-                                    Text(
-                                      "Schedule: ${post.postDate}",
-                                      style: textstyle,
-                                    ),
-                                    Text(
-                                      post.postPlatform,
-                                      style: textstyle,
-                                    ),
-                                  ],
-                                ),
-                                Text(
-                                  post.postTime,
-                                  style: textstyle,
-                                ),
-                              ],
-                            ),
+                        ),
+                        Padding(
+                          padding: const EdgeInsets.symmetric(horizontal: 6),
+                          child: Row(
+                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                            crossAxisAlignment: CrossAxisAlignment.start,
+                            children: [
+                              Column(
+                                mainAxisAlignment: MainAxisAlignment.start,
+                                crossAxisAlignment: CrossAxisAlignment.start,
+                                children: [
+                                  Text(
+                                    "Schedule: ${post.postDate}",
+                                    style: textstyle,
+                                  ),
+                                  Text(
+                                    post.postPlatform,
+                                    style: textstyle,
+                                  ),
+                                ],
+                              ),
+                              Text(
+                                post.postTime,
+                                style: textstyle,
+                              ),
+                            ],
                           ),
-                        ],
-                      ),
+                        ),
+                      ],
                     ),
                   ),
                 ),
-              );
-            },
-          ).toList(),
+              ),
+            );
+          }).toList(),
         ),
       ],
     );
